move VerticalFlexbox to components folder
